Add linting and code coverage
