+++ README.rst
@@ -21,6 +21,55 @@ To get started developing against Koku first clone a local copy of the git repos
 
     git clone https://github.com/quipucords/yupana
 
+Developing inside a virtual environment is recommended. A Pipfile is provided. Pipenv is recommended for combining virtual environment (virtualenv) and dependency management (pip). To install pipenv, use pip ::
+
+    pip3 install pipenv
+
+Then project dependencies and a virtual environment can be created using ::
+
+    pipenv install --dev
+
+To activate the virtual environment run ::
+
+    pipenv shell
+
+Configuration
+^^^^^^^^^^^^^
+
+This project is developed using the Django web framework. Many configuration settings can be read in from a `.env` file. An example file `.env.example` is provided in the repository. To use the defaults simply ::
+
+    cp .env.example .env
+
+
+Modify as you see fit.
+
+
+Server
+^^^^^^
+
+To run a local dev Django server you can use ::
+
+    make run-migrations
+    make serve
+
+Testing and Linting
+-------------------
+
+Koku uses tox to standardize the environment used when running tests. Essentially, tox manages its own virtual environment and a copy of required dependencies to run tests. To ensure a clean tox environment run ::
+
+    tox -r
+
+This will rebuild the tox virtual env and then run all tests.
+
+To run unit tests specifically::
+
+    tox -e py36
+
+To lint the code base ::
+
+    tox -e lint
+
+
 .. |license| image:: https://img.shields.io/github/license/quipucords/yupana.svg
 .. |Updates| image:: https://pyup.io/repos/github/quipucords/yupana/shield.svg
    :target: https://pyup.io/repos/github/quipucords/yupana/
