+++ tox.ini
@@ -1,0 +1,48 @@
+[tox]
+envlist = py36, lint
+skipsdist = True
+
+[flake8]
+; D106 = Missing docstring in public nested class
+; D212 = Multi-line docstring summary should start at the first line
+ignore = D106,D212
+max-complexity = 10
+max-line-length = 120
+exclude =
+  **/__init__.py
+  **/manage.py
+  **/settings/*.py
+  **/apps.py
+  **/migrations/*.py
+  **/.tox/*.py
+import-order-style = pycharm
+application-import-names = config, api
+
+[testenv]
+passenv = CI TRAVIS TRAVIS_*
+
+deps =
+  pipenv
+  codecov
+commands =
+  pipenv install --dev
+  coverage run {toxinidir}/yupana/manage.py test -v 2 {posargs: api}
+  coverage report --show-missing
+
+[testenv:lint]
+deps =
+  flake8
+  flake8-docstrings
+  flake8-import-order
+  flake8-quotes
+  pipenv
+  pylint
+  pylint-plugin-utils
+  pylint-django
+setenv =
+  DJANGO_SETTINGS_MODULE=config.settings.test
+  PYTHONPATH={toxinidir}/yupana
+commands =
+  flake8 yupana
+  pipenv install --dev --ignore-pipfile
+  pylint --load-plugins=pylint_django yupana/config
