+++ yupana/api/admin.py
@@ -14,7 +14,5 @@
 # You should have received a copy of the GNU Affero General Public License
 # along with this program.  If not, see <https://www.gnu.org/licenses/>.
 #
-
-from django.contrib import admin
-
-# Register your models here.
+"""Admin module for Django server application."""
+# from django.contrib import admin
