+++ yupana/api/urls.py
@@ -15,7 +15,6 @@
 #
 """Describes the urls and patterns for the API application."""
 from django.conf.urls import include, url
-from rest_framework.authtoken import views
 from rest_framework.routers import DefaultRouter
 
 from api.views import (status)
